backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4
django-cors-headers>=4.4,<5
psycopg[binary]>=3.2,<4
stripe>=10.0,<11
django-filter>=24.0,<25
python-dotenv>=1.0,<2
gunicorn>=22.0,<23
whitenoise>=6.7,<7
Pillow>=10.4,<11
